Changes `m_owners` and `m_ownerIndex` to use address instead of uint.
It is unclear why this code was using `uint` previously instead of `address`.  If `address` is what is intended, `address` should be used.  Using `address` not only makes the intent more clear, it allows the static analyzer to validate the code.

Also, tooling will now work better with this contract since the types are correct.

If these variables were intentionally `uint`, I am curious to know the reason why and at the least there should be a comment in the code indicating why they are `uint`.